backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
faker==20.0.0
